+++ src/main/java/com/csl/adi/repository/OrigemRepository.java
@@ -2,7 +2,7 @@ package com.csl.adi.repository;
 
 import com.csl.adi.model.Origem;
 
-public interface OrigemRepository extends BaseRepository<Origem, Integer> {
+public interface OrigemRepository extends BaseRepository<Origem, Long> {
 
     Origem findBySigla(String sigla);
