+++ src/main/java/com/csl/adi/repository/UrlRepository.java
@@ -4,12 +4,12 @@ import com.csl.adi.model.Url;
 
 import java.util.List;
 
-public interface UrlRepository extends BaseRepository<Url, Integer> {
+public interface UrlRepository extends BaseRepository<Url, Long> {
 
-    List<Url> findByOrigem (Integer origem);
+    List<Url> findByOrigem (Long origem);
 
     Url findByDescricao (String descricao);
 
-    Url findByOrigemAndDescricao (Integer origem, String descricao);
+    Url findByOrigemAndDescricao (Long origem, String descricao);
 
 }
